Test the activate script from tox and travis
Wrap the test for the activate script with tox. While the test could be
run standalone without any extra dependencies, it seems like a
convenience to have everything wrapped by tox. It also makes it easier
to plug this test into travis CI.

The downside of this is that we now run a virtualenv inside another
virtualenv. That imposes some extra output which has to be accounted for
in the test assertion.